Cleanup

+++ src/ModifyRespawnParty/MobilePartyHelperPatch.cs
@@ -23,7 +23,7 @@ namespace CombatModCollection
             ++hero.NumberOfCreatedParties;
             spawnedParty.AddElementToMemberRoster(hero.CharacterObject, 1, true);
             int troopNumberLimit = hero == Hero.MainHero || hero.Clan == Clan.PlayerClan ?
-                (int)Settings.Instance.Strategy_ModifyRespawnParty_PlayerPartySizeOnRespawn : (int)Settings.Instance.Strategy_ModifyRespawnParty_AILordPartySizeOnRespawn;
+                Settings.Instance.Strategy_ModifyRespawnParty_PlayerPartySizeOnRespawn : Settings.Instance.Strategy_ModifyRespawnParty_AILordPartySizeOnRespawn;
             if (!Campaign.Current.GameStarted)
             {
                 float num = (float)(1.0 - (double)MBRandom.RandomFloat * (double)MBRandom.RandomFloat);
